% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classAnalysisPage.R
\name{classAnalysisPageUI}
\alias{classAnalysisPageUI}
\title{UI of class analysis page}
\usage{
classAnalysisPageUI(id)
}
\arguments{
\item{id}{Namespace of this page.}
}
\value{
Same to \code{shiny::fluidRow} function.
}
\description{
UI of class analysis page
}
