% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventAnalysisPage.R
\name{eventAnalysisPage}
\alias{eventAnalysisPage}
\title{Server logic of event analysis page}
\usage{
eventAnalysisPage(input, output, session, df)
}
\arguments{
\item{input, output, session}{Arguments passed by \code{shiny::callModule} .}

\item{df}{A data frame.}
}
\description{
Server logic of event analysis page
}
