% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventAnalysisPage.R
\name{eventAnalysisPageUI}
\alias{eventAnalysisPageUI}
\title{Create event analysis page UI}
\usage{
eventAnalysisPageUI(id)
}
\arguments{
\item{id}{Namespace for the page.}
}
\description{
Create event analysis page UI
}
