% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info.R
\name{get_local_users}
\alias{get_local_users}
\title{Get local user list}
\usage{
get_local_users(wiznote_home)
}
\arguments{
\item{wiznote_home}{The home directory of WizNote program.}
}
\value{
A list
}
\description{
Get local user list
}
