% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{html_table_to_dataframe}
\alias{html_table_to_dataframe}
\title{Convert html table element to data.frame}
\usage{
html_table_to_dataframe(x, header = NA, trim = TRUE, fill = FALSE, dec = ".")
}
\arguments{
\item{x}{A node, node set or document.}

\item{header}{Use first row as header? If `NA`, will use first row
if it consists of `<th>` tags.}

\item{trim}{Remove leading and trailing whitespace within each cell?}

\item{fill}{If `TRUE`, automatically fill rows with fewer than
the maximum number of columns with `NA`s.}

\item{dec}{The character used as decimal mark.}
}
\value{
A data.frame
}
\description{
Convert html table element to data.frame
}
