% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{mainServer}
\alias{mainServer}
\title{Main server logic}
\usage{
mainServer(input, output, session)
}
\arguments{
\item{input, output}{Arguments passed by \code{shiny::callModule} .}
}
\description{
Main server logic
}
