% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info.R
\name{new_user_info}
\alias{new_user_info}
\title{Construct User information object.}
\usage{
new_user_info(user_name, data_location, wiznote_home)
}
\arguments{
\item{user_name}{A string, the email account of WizNote}

\item{data_location}{A string, where the documents of daily recods located.}

\item{wiznote_home}{The home directory of WizNote program.}
}
\value{
An `user_info`` S3 object.
}
\description{
`new_user_info` return an object that contains WizNote account
 information.
}
