% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_time.R
\name{plot_class_time}
\alias{plot_class_time}
\title{Create a column plot based on the time cost of each event class.}
\usage{
plot_class_time(df, cls = NULL, time_units = "hours", ...)
}
\arguments{
\item{df}{A tidy tibble}

\item{cls}{String vector contains which classes you want to plot.}

\item{time_units}{Unit of time cost, such as 'hours', 'minutes' and so on.}

\item{...}{Arguments will be passed to \code{ggplot2::geom_col} .}
}
\value{
A ggplot object
}
\description{
Create a column plot based on the time cost of each event class.
}
