% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{query_page_metainfo}
\alias{query_page_metainfo}
\title{Find document meta information from WizNote database}
\usage{
query_page_metainfo(user_info, start, end)
}
\arguments{
\item{user_info}{An user information S3 object.}

\item{start, end}{Any input for `lubridate::ymd` function.}
}
\value{
A tibble that contains meta informatin of WizNote document
}
\description{
`query_page_metainfo` will extract document meta information from WizNote
database with the given start to end date.
}
