% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_records.R
\name{query_records}
\alias{query_records}
\title{Extract records from given page document}
\usage{
query_records(user_info, start, end)
}
\arguments{
\item{user_info}{An user information S3 object.}

\item{date}{Any input for `lubridate::ymd` function. One page of time ledge
per day.}
}
\value{
A tidy tibble that contains event time records of your each day.
}
\description{
`query_records` will produce tidy data from WizNote database.
}
