% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{read_html_from_doc}
\alias{read_html_from_doc}
\title{Get html text from WizNotePlus document zipfile}
\usage{
read_html_from_doc(filename)
}
\arguments{
\item{filename}{WizNote document zipfile name}
}
\value{
string. The html text of document zipfile.
}
\description{
Get html text from WizNotePlus document zipfile
}
