% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table_from_doc.R
\name{read_table_from_doc}
\alias{read_table_from_doc}
\title{Extract table from html text}
\usage{
read_table_from_doc(user_info, doc_guid)
}
\arguments{
\item{user_info}{An user information S3 object.}

\item{doc_guid}{A string. The GUID of WizNote document.}
}
\value{
A tibble that contains event time records.
}
\description{
Extract table from html text
}
