% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPage.R
\name{summaryPage}
\alias{summaryPage}
\title{Server logic of histogram page}
\usage{
summaryPage(input, output, session, df)
}
\arguments{
\item{input, output, session}{Arguments passed by \code{shiny::callModule} .}

\item{df}{A data frame which used to plot histogram.}
}
\description{
Server logic of histogram page
}
