% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_info.R
\name{user_info}
\alias{user_info}
\title{Helper to construct user info object.}
\usage{
user_info(user_name = "", data_location = "/My Notes/")
}
\arguments{
\item{user_name}{A string, the email account of WizNote}

\item{data_location}{A string, where the documents of daily recods located.}
}
\value{
An user information S3 object.
}
\description{
Helper to construct user info object.
}
