% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_input.R
\name{tidy_records}
\alias{tidy_records}
\title{Organize records table}
\usage{
tidy_records(records)
}
\arguments{
\item{records}{A tibble.}

\item{date}{Any input for `lubridate::ymd` function.}
}
\value{
A tidy tibble.
}
\description{
`tidy_records` is very verbose.
}
